postgresql connection

+++ nodejs/notes/13-Models_Todo_2/todo.model.js
@@ -5,8 +5,12 @@
 // MODELS:
 
 const { Sequelize, DataTypes } = require('sequelize')
-// sequelize instant oluştur
+// sequelize instance oluştur
 const sequelize = new Sequelize('sqlite:./db.sqlite3')
+
+// POSTGRESQL CONNECTİON:
+// $ npm i pg pg-hstore
+// const sequelize = new Sequelize('postgres://userCH15:12345678@localhost:5432/testCH15')
 
 // define methodu sequelize modeli oluştur:
 // her bir model veritabanında bir tabloya denk gelir
@@ -32,7 +36,8 @@ const Todo = sequelize.define('todos', {
     description: DataTypes.TEXT, // ShortHand Using.
 
     priority: { // -1: Low, 0: Norm, 1: High
-        type: DataTypes.TINYINT,
+        // type: DataTypes.TINYINT, // postgresql desteklemiyor.
+        type: DataTypes.INTEGER,
         allowNull: false,
         defaultValue: 0
     },
@@ -49,7 +54,7 @@ const Todo = sequelize.define('todos', {
 })
  // Syncronization
  // Model bilgilerini db  ye uygula
- //  sequelize.sync()  // CREATE TABLE
+//   sequelize.sync()  // CREATE TABLE
  // sequelize.sync({ force: true }) // DROP TABLE & CREATE TABLE
  // sequelize.sync({ alter: true }) // TO BACKUP & DROP TABLE & CREATE TABLE & FROM BACKUP
